requirements.txt
flask
xgboost
gradio_client
flask-cors
numpy
scikit-learn
prophet
pytrends
python-dotenv
openai